pyproject.toml
[project]
name = "a2a-python-ui"
version = "0.1.0"
description = "Agent2Agent UI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "mesop>=1.0.0",
    "a2a-sdk>=0.3.0",
    "pandas>=2.2.0",
    "google-genai>=1.9.0",
    "google-adk[a2a]>=1.7.0",
    "litellm",
]

[tool.hatch.build.targets.wheel]
packages = ["a2a_ui"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
a2a_ui = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.11.2"]
